feat(isotp): implement multi-frame transfer with flow control and timers
Add First Frame, Consecutive Frame and Flow Control handling on top of
the existing Single Frame support, with reassembly, sequence number
checking and the N_Bs / N_Cr timers. Reception and transmission are
explicit state machines in isotp_rx.c and isotp_tx.c rather than nested
conditionals, so every transition is visible and an unforeseen case
falls through to a defined reset instead of a half-open context.

Time is passed in as a parameter; the protocol layers never read a
clock. Timer tests therefore advance the clock by hand and run
instantly, and a microcontroller port needs no POSIX clock. Elapsed time
uses unsigned subtraction, which stays correct across the 32-bit
millisecond wraparound at 49 days - there is a test for that too.

A First Frame announcing more than the reassembly buffer is answered
with FlowStatus Overflow and reserves nothing, rather than being
silently dropped or trusted.

Multi-frame gave the ECU and the tester an identical, non-trivial
service loop, which is the trigger recorded in decision D2 for building
the transport layer. src/platform now holds the raw CAN socket and the
ISO-TP session loop; both applications shrank to their actual job, and
the CAN identifier is now masked against its flag bits instead of being
compared raw.

UDS response lengths widen to uint16_t: the 7-byte ceiling was a
transport limitation that no longer exists. The VIN is readable end to
end, and the bus shows the textbook sequence 10 14 / 30 00 00 / 21 / 22.

+++ src/ecu/ecu_data.c
@@ -87,8 +87,8 @@ void ecu_data_tick(ecu_data_t *data)
 static uds_result_t emit_bytes(const uint8_t *src,
                                uint8_t src_len,
                                uint8_t *out,
-                               uint8_t out_capacity,
-                               uint8_t *out_len)
+                               uint16_t out_capacity,
+                               uint16_t *out_len)
 {
     uint8_t i;
 
@@ -109,8 +109,8 @@ static uds_result_t emit_bytes(const uint8_t *src,
 /* Entier 16 bits, poids fort en premier (convention UDS). */
 static uds_result_t emit_u16(uint16_t value,
                              uint8_t *out,
-                             uint8_t out_capacity,
-                             uint8_t *out_len)
+                             uint16_t out_capacity,
+                             uint16_t *out_len)
 {
     uint8_t buf[2];
 
@@ -122,8 +122,8 @@ static uds_result_t emit_u16(uint16_t value,
 
 uds_result_t ecu_data_read_did(uint16_t did,
                                uint8_t *out,
-                               uint8_t out_capacity,
-                               uint8_t *out_len,
+                               uint16_t out_capacity,
+                               uint16_t *out_len,
                                void *user_ctx)
 {
     const ecu_data_t *data = (const ecu_data_t *)user_ctx;
